Rename to delete_user
