+++ lib/features/settings/setting_account_link/page.dart
@@ -11,7 +11,7 @@ import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
 import 'package:pilll/components/atoms/text_color.dart';
 import 'package:pilll/components/page/discard_dialog.dart';
-import 'package:pilll/features/settings/setting_account_link/components/delete_user_button.dart';
+import 'package:pilll/features/settings/setting_account_link/components/delete_user.dart';
 import 'package:pilll/entity/link_account_type.dart';
 import 'package:pilll/features/error/error_alert.dart';
 import 'package:pilll/features/error/page.dart';
